feat: 내비게이션바 클릭된 내용에 따라 렌더링 다르게 (#13)
