+++ src/components/group/Food.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Food = () => {
+  return (
+    <div>Food</div>
+  )
+}
+
+export default Food
