+++ src/components/group/Setting.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Setting = () => {
+  return (
+    <div>Setting</div>
+  )
+}
+
+export default Setting
